+++ caddisfly-app/app/src/androidTest/java/org/akvo/caddisfly/EspressoTest.java
@@ -17,8 +17,11 @@
 package org.akvo.caddisfly;
 
 import android.app.Activity;
+import android.content.SharedPreferences;
 import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
+import android.os.Environment;
+import android.preference.PreferenceManager;
 import android.support.test.espresso.Espresso;
 import android.support.test.espresso.NoActivityResumedException;
 import android.support.test.espresso.core.deps.guava.collect.Iterables;
@@ -26,9 +29,17 @@ import android.support.test.espresso.matcher.BoundedMatcher;
 import android.support.test.internal.util.Checks;
 import android.support.test.runner.lifecycle.ActivityLifecycleMonitorRegistry;
 import android.support.test.runner.lifecycle.Stage;
+import android.support.test.uiautomator.By;
+import android.support.test.uiautomator.UiDevice;
+import android.support.test.uiautomator.UiObject;
+import android.support.test.uiautomator.UiObject2;
+import android.support.test.uiautomator.UiObjectNotFoundException;
+import android.support.test.uiautomator.UiScrollable;
+import android.support.test.uiautomator.UiSelector;
 import android.test.ActivityInstrumentationTestCase2;
 import android.view.View;
 import android.widget.Button;
+import android.widget.TextView;
 
 import com.squareup.spoon.Spoon;
 
@@ -37,6 +48,7 @@ import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
 
+import java.io.File;
 import java.text.DecimalFormatSymbols;
 import java.util.HashMap;
 
@@ -56,12 +68,13 @@ public class EspressoTest
         extends ActivityInstrumentationTestCase2<MainActivity> {
 
     static final String mCurrentLanguage = "en";
-    static boolean mTakeScreenshots = true;
+    static boolean mTakeScreenshots = false;
     static int mCounter = 0;
-
+    protected String appName = "Akvo Caddisfly";
     HashMap<String, String> stringHashMapEN = new HashMap<>();
     HashMap<String, String> stringHashMapFR = new HashMap<>();
     HashMap<String, String> currentHashMap;
+    private UiDevice mDevice;
 
     public EspressoTest() {
         super(MainActivity.class);
@@ -120,6 +133,10 @@ public class EspressoTest
         }
 
         getActivity();
+
+        // Initialize UiDevice instance
+        mDevice = UiDevice.getInstance(getInstrumentation());
+
     }
 
     @SuppressWarnings("EmptyMethod")
@@ -127,33 +144,294 @@ public class EspressoTest
     }
 
     public void test001_Language() {
-        onView(withId(R.id.action_settings))
-                .perform(click());
 
-        onView(withText(R.string.language))
-                .perform(click());
+        onView(withId(R.id.action_settings)).perform(click());
+
+        onView(withText(R.string.about)).check(matches(isDisplayed())).perform(click());
+
+        enterDiagnosticMode();
+
+        goToMainScreen();
+
+        onView(withId(R.id.disableDiagnosticButton)).check(matches(isDisplayed()));
+
+        leaveDiagnosticMode();
+
+        onView(withId(R.id.action_settings)).perform(click());
+
+        onView(withText(R.string.language)).perform(click());
 
         onData(hasToString(startsWith(currentHashMap.get("language")))).perform(click());
     }
 
-    public void test002_About() {
+    public void test002_Screenshots() {
+
+        SharedPreferences prefs =
+                PreferenceManager.getDefaultSharedPreferences(this.getInstrumentation().getTargetContext());
+        prefs.edit().clear().apply();
+
+        String path = Environment.getExternalStorageDirectory().getPath() + "/org.akvo.caddisfly/screenshots";
+
+        File folder = new File(path);
+        if (!folder.exists()) {
+            //noinspection ResultOfMethodCallIgnored
+            folder.mkdirs();
+        }
+
+        testLanguageReset();
+
+        mDevice.waitForWindowUpdate("", 2000);
+
+        goToMainScreen();
 
         final Button button = (Button) getActivity().findViewById(R.id.surveyButton);
         assertNotNull(button);
 
+        //Main Screen
         takeScreenshot();
 
-        onView(withId(R.id.action_settings))
-                .perform(click());
+        onView(withId(R.id.action_settings)).perform(click());
 
+        //Settings Screen
         takeScreenshot();
 
         onView(withText(R.string.about)).check(matches(isDisplayed())).perform(click());
 
+        //About Screen
+        takeScreenshot();
+
+        enterDiagnosticMode();
+
+        Espresso.pressBack();
+
+        onView(withText(R.string.calibrate)).perform(click());
+
+        onView(withText(currentHashMap.get("fluoride"))).perform(click());
+
+        onView(withId(R.id.menu_load)).perform(click());
+
+        onView(withText("Test")).perform(click());
+
+        goToMainScreen();
+
+        leaveDiagnosticMode();
+
+        try {
+            Thread.sleep(4000);
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        }
+
+        onView(withId(R.id.action_settings)).perform(click());
+
+        onView(withText(R.string.language)).perform(click());
+
+        mDevice.waitForWindowUpdate("", 1000);
+
+        //Language Dialog
+        takeScreenshot();
+
+        onView(withId(android.R.id.button2)).perform(click());
+
+        onView(withText(R.string.updateCheck)).perform(click());
+
+        onView(withText(R.string.dataConnection)).check(matches(isDisplayed()));
+        onView(withText(R.string.enableInternet)).check(matches(isDisplayed()));
+
+        mDevice.waitForWindowUpdate("", 1000);
+
+        //Enable Internet Dialog
         takeScreenshot();
 
         Espresso.pressBack();
 
+        onView(withText(R.string.calibrate)).perform(click());
+
+        //Test Types Screen
+        takeScreenshot();
+
+        onView(withText(currentHashMap.get("fluoride"))).perform(click());
+
+        //Ranges Screen
+        takeScreenshot();
+
+        DecimalFormatSymbols dfs = new DecimalFormatSymbols();
+        onView(withText("2" + dfs.getDecimalSeparator() + "00 ppm")).perform(click());
+
+        //Test Start Screen
+        takeScreenshot();
+
+        onView(withId(R.id.startButton)).perform(click());
+
+        //Test Progress Screen
+        takeScreenshot();
+
+        Espresso.pressBack();
+
+        Espresso.pressBack();
+
+        Espresso.pressBack();
+
+        onView(withText(R.string.calibrate)).perform(click());
+
+        onView(withText(currentHashMap.get("electricalConductivity"))).perform(click());
+
+        //Calibrate EC
+        takeScreenshot();
+
+        onView(withId(R.id.startButton)).perform(click());
+
+        //EC not found dialog
+        takeScreenshot();
+
+        onView(withId(android.R.id.button2)).perform(click());
+
+        goToMainScreen();
+
+        onView(withId(R.id.surveyButton)).check(matches(isClickable()));
+
+        onView(withId(R.id.surveyButton)).perform(click());
+
+        mDevice.waitForWindowUpdate("", 2000);
+
+        clickListViewItem("Automated Tests");
+
+        clickListViewItem("CREATE NEW DATA POINT");
+
+        clickListViewItem("All Tests");
+
+        try {
+
+            mDevice.findObject(new UiSelector().text("Use External Source")).click();
+
+            mDevice.waitForWindowUpdate("", 2000);
+
+        } catch (UiObjectNotFoundException e) {
+            e.printStackTrace();
+        }
+
+        onView(withId(R.id.noDilutionButton)).check(matches(isDisplayed()));
+
+        //Dilution dialog
+        takeScreenshot();
+
+        mDevice.pressBack();
+        mDevice.pressBack();
+        mDevice.pressBack();
+        mDevice.pressBack();
+        mDevice.pressBack();
+
+        startApp();
+
+        onView(withId(R.id.surveyButton)).check(matches(isClickable()));
+
+        onView(withId(R.id.surveyButton)).perform(click());
+
+        mDevice.waitForWindowUpdate("", 2000);
+
+        clickListViewItem("Automated Tests");
+
+        clickListViewItem("CREATE NEW DATA POINT");
+
+        clickListViewItem("All Tests");
+
+        try {
+            mDevice.findObject(new UiSelector().text("Next")).click();
+
+            mDevice.waitForWindowUpdate("", 2000);
+
+            mDevice.findObject(new UiSelector().text("Use External Source")).click();
+
+            mDevice.waitForWindowUpdate("", 2000);
+
+        } catch (UiObjectNotFoundException e) {
+            e.printStackTrace();
+        }
+
+        //Connect EC Sensor Screen
+        takeScreenshot();
+
+        onView(withId(android.R.id.button2)).perform(click());
+
+        try {
+            mDevice.findObject(new UiSelector().text("Next")).click();
+
+            mDevice.waitForWindowUpdate("", 2000);
+
+            mDevice.findObject(new UiSelector().text("Use External Source")).click();
+
+            mDevice.waitForWindowUpdate("", 2000);
+
+        } catch (UiObjectNotFoundException e) {
+            e.printStackTrace();
+        }
+
+        takeScreenshot();
+
+        onView(withId(R.id.backButton)).perform(click());
+
+        try {
+            mDevice.findObject(new UiSelector().text("Next")).click();
+
+            mDevice.waitForWindowUpdate("", 2000);
+
+            mDevice.findObject(new UiSelector().text("Use External Source")).click();
+
+            mDevice.waitForWindowUpdate("", 2000);
+
+        } catch (UiObjectNotFoundException e) {
+            e.printStackTrace();
+        }
+
+        takeScreenshot();
+
+        onView(withId(android.R.id.button1)).perform(click());
+
+        mDevice.pressBack();
+        mDevice.pressBack();
+        mDevice.pressBack();
+        mDevice.pressBack();
+
+
+    }
+
+    private void startApp() {
+        // Start from the home screen
+        mDevice.pressHome();
+        mDevice.waitForWindowUpdate("", 2000);
+        UiObject2 allAppsButton = mDevice.findObject(By.desc("Apps"));
+        allAppsButton.click();
+        mDevice.waitForWindowUpdate("", 2000);
+
+        UiScrollable appViews = new UiScrollable(new UiSelector().scrollable(true));
+        appViews.setAsHorizontalList();
+
+        UiObject settingsApp = null;
+        try {
+            settingsApp = appViews.getChildByText(new UiSelector().className(TextView.class.getName()), appName);
+        } catch (UiObjectNotFoundException e) {
+            e.printStackTrace();
+        }
+        try {
+            if (settingsApp != null) {
+                settingsApp.clickAndWaitForNewWindow();
+            }
+        } catch (UiObjectNotFoundException e) {
+            e.printStackTrace();
+        }
+
+        mDevice.waitForWindowUpdate("", 2000);
+
+        assertTrue("Unable to detect app", settingsApp != null);
+    }
+
+    private void goToMainScreen() {
+
+        Button button = (Button) getCurrentActivity().findViewById(R.id.surveyButton);
+        while (button == null) {
+            Espresso.pressBack();
+            button = (Button) getCurrentActivity().findViewById(R.id.surveyButton);
+        }
     }
 
     public void testCalibrateSwatches() {
@@ -190,16 +468,10 @@ public class EspressoTest
 
         onView(withText(R.string.calibrate)).perform(click());
 
-        takeScreenshot();
-
         onView(withText(currentHashMap.get("fluoride"))).perform(click());
-
-        takeScreenshot();
 
         DecimalFormatSymbols dfs = new DecimalFormatSymbols();
         onView(withText("0" + dfs.getDecimalSeparator() + "00 ppm")).perform(click());
-
-        takeScreenshot();
 
         Espresso.pressBack();
 
@@ -212,28 +484,6 @@ public class EspressoTest
         Espresso.pressBack();
 
         Espresso.pressBack();
-
-//        onView(withText("Nitrite")).perform(click());
-//
-//        onView(withText("3.00 ppm")).perform(click());
-//
-//        Espresso.pressBack();
-//
-//        Espresso.pressBack();
-//
-//        onView(withText("pH")).perform(click());
-//
-//        onView(withText("9.00 pH")).perform(click());
-//
-//        onView(withId(R.id.startButton)).perform(click());
-
-        //onView(withText(R.string.selectDilution)).check(matches(isDisplayed()));
-
-        //onView(withId(android.R.id.button2)).perform(click());
-
-//        Espresso.pressBack();
-//
-//        Espresso.pressBack();
 
         Espresso.pressBack();
     }
@@ -274,19 +524,31 @@ public class EspressoTest
 
         onView(withId(R.id.surveyButton)).perform(click());
 
+        mDevice.waitForWindowUpdate("", 2000);
+
+        clickListViewItem("Automated Tests");
+
+        clickListViewItem("CREATE NEW DATA POINT");
+
+        clickListViewItem("All Tests");
+
         try {
-            Thread.sleep(10000);
-        } catch (InterruptedException e) {
+
+            mDevice.findObject(new UiSelector().text("Use External Source")).click();
+
+            mDevice.waitForWindowUpdate("", 2000);
+
+        } catch (UiObjectNotFoundException e) {
             e.printStackTrace();
         }
 
-        onView(withId(R.id.hundredPercentButton)).check(matches(isDisplayed()));
+        onView(withId(R.id.noDilutionButton)).check(matches(isDisplayed()));
 
-        onView(withId(R.id.fiftyPercentButton)).check(matches(isDisplayed()));
+        onView(withId(R.id.percentButton1)).check(matches(isDisplayed()));
 
-        onView(withId(R.id.twentyFivePercentButton)).check(matches(isDisplayed()));
+        onView(withId(R.id.percentButton2)).check(matches(isDisplayed()));
 
-        onView(withId(R.id.hundredPercentButton)).perform(click());
+        onView(withId(R.id.noDilutionButton)).perform(click());
 
         onView(withId(R.id.startButton)).perform(click());
 
@@ -296,6 +558,11 @@ public class EspressoTest
             Espresso.pressBack();
         } catch (NoActivityResumedException ignored) {
         }
+
+        mDevice.pressBack();
+        mDevice.pressBack();
+        mDevice.pressBack();
+        mDevice.pressBack();
     }
 
     public void testCalibrateSensor() {
@@ -471,7 +738,16 @@ public class EspressoTest
     }
 
 
-    public void testLeaveDiagnosticMode() {
+    private void enterDiagnosticMode() {
+
+        for (int i = 0; i < 10; i++) {
+            onView(withId(R.id.textVersion)).perform(click());
+        }
+
+    }
+
+
+    private void leaveDiagnosticMode() {
         onView(withId(R.id.disableDiagnosticButton)).perform(click());
 
         onView(withId(R.id.disableDiagnosticButton)).check(matches(not(isDisplayed())));
@@ -547,38 +823,43 @@ public class EspressoTest
         return activity[0];
     }
 
+    private void takeSpoonScreenshot() {
+        if (mTakeScreenshots) {
+            Spoon.screenshot(getCurrentActivity(), "screen-" + mCounter++ + "-" + mCurrentLanguage);
+        }
+    }
+
     private void takeScreenshot() {
         if (mTakeScreenshots) {
-            //int SDK_VERSION = android.os.Build.VERSION.SDK_INT;
-            //if (SDK_VERSION >= 17) {
-//            try {
-//                Thread.sleep(500);
-//            } catch (InterruptedException e) {
-//                e.printStackTrace();
-//            }
+            int SDK_VERSION = android.os.Build.VERSION.SDK_INT;
+            if (SDK_VERSION >= 17) {
+                try {
+                    Thread.sleep(500);
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
 
-//            File path = new File(Environment.getExternalStorageDirectory().getPath() +
-//                    "/org.akvo.caddisfly/screen-" + mCounter++ + "-" + mCurrentLanguage + ".png");
-
-            Spoon.screenshot(getCurrentActivity(), "screen-" + mCounter++ + "-" + mCurrentLanguage);
-            //}
+                File path = new File(Environment.getExternalStorageDirectory().getPath() +
+                        "/org.akvo.caddisfly/screenshots/screen-" + mCounter++ + "-" + mCurrentLanguage + ".png");
+                mDevice.takeScreenshot(path, 0.5f, 60);
+            }
         }
     }
 
 
-//    public void tearDown() throws Exception {
-//        Thread.sleep(1000);
-//        goBackN();
-//        super.tearDown();
-//    }
-//
-//    private void goBackN() {
-//        final int N = 3; // how many times to hit back button
-//        try {
-//            for (int i = 0; i < N; i++)
-//                Espresso.pressBack();
-//        } catch (NoActivityResumedException e) {
-//            //Log.e(TAG, "Closed all activities", e);
-//        }
-//    }
+    private void clickListViewItem(String name) {
+        UiScrollable listView = new UiScrollable(new UiSelector());
+        listView.setMaxSearchSwipes(100);
+        listView.waitForExists(5000);
+        UiObject listViewItem = null;
+        try {
+            listView.scrollTextIntoView(name);
+            listViewItem = listView.getChildByText(new UiSelector()
+                    .className(android.widget.TextView.class.getName()), "" + name + "");
+            listViewItem.click();
+        } catch (UiObjectNotFoundException e) {
+            e.printStackTrace();
+        }
+        System.out.println("\"" + name + "\" ListView item was clicked.");
+    }
 }